+++ Assets/Prefabs/UI_Prefabs/UpgradeMenuCanvas.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b5e190e8c08e374ca39ffe40aa953d2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
