--- Assets/Scripts/CreateGrid.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63884f73dc95d0644b8a28595c2c3a40
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
