--- Assets/Scripts/Player/Spawn.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c80b55d2af6d5fe47b70a09573768ee4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
