--- Assets/Scripts/Player/Spawn/Grid.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2583d962be3e98409812963efa22e1d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
