--- Assets/Scripts/Player/Spawn/SpawnManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d5e8718691edf9d4a9c66044b85dd29b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
