--- Assets/Scripts/Player/Spawn/TileCollision.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff6254a711bacf642973862d7e183ca1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
