+++ Assets/Scripts/Spawn.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 294328ebb7f9da64ebf3929c684ebbb5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
