--- Assets/Scripts/Tile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c50745c40a87a98449f22f3349a1552e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
