+++ Assets/SwitchBetweenRoundModeandUpgradeMode.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// Controls Actions When The Player Switches To Round Mode or Upgrade Mode
+
+public class SwitchBetweenRoundModeandUpgradeMode : MonoBehaviour
+{
+    // Actions That Occur When A Player Switches To Upgrade Mode
+    public void SwitchToUpgradeMode() {
+
+    }
+
+    // Actions That Occur When A Player Switches To Round Mode
+    public void SwitchToRoundMode() {
+        
+    }
+}
